Fix backspace and overflow bugs
Backspace used to delete the entire results on
screen for intermediate calculations. This is
because results were stored as single strings.
Checks are added to lookout for this and preform
manipulation of the textContent instead. Spread
operator applied to screen to split the string.

Overflow errors are not checked for, now added to
catch and handle these errors with
showErrorDisplay() function.